+++ CHANGELOG.md
@@ -1,10 +1,13 @@
 # Changelog
-##0.8.0#
+##0.8.0
 ### Breaking API Changes
 * removed deprecated code
+* `Aspect` no longer uses static methods, use named constructors instead
+(migration: replace `Aspect.getAspectF` with`new Aspect.f`)
+* methods in `Aspect` no longer return the aspect, use cascading operator to chain calls
 ### Enhancements
 * `world.destroy()` for cleaning up `EntitySystem`s and `Manager`s
-##0.7.0#
+##0.7.0
 ### Breaking API Changes
 * renamed `Poolable` to `Pooled`
 * renamed `ComponentPoolable` to `PooledComponent`